Move languages into subfolder
