+++ src/js/languages/es6.js
