+++ src/js/languages/swift.js
